Fix crash in Liquid::C::BlockBody
When Liquid::C::BlockBody#freeze is called, it calls
document_body_write_block_body which sets the document_body_entry but
does not yet set the compiled flag to true. During this time, the only
reference to the Liquid::C::DocumentBody object is in the instance
variables of the parse_context which is marked movable by Ruby. This
causes the self reference in document_body_t to be moved by compaction
causing it to point to an incorrect object.

Co-Authored-By: Matt Valentine-House <matt@eightbitraptor.com>