Fix crash in vm_assembler_pool_gc_mark with compaction
This commit adds a marking function to Liquid::C::VMAssemblerPool.

In some cases, the only reference to the Liquid::C::VMAssemblerPool
object is in the instance variables of the Liquid::ParseContext which is
marked movable by Ruby. This causes the self reference in
vm_assembler_pool_t to be moved by compaction causing it to point to an
incorrect object.